+++ mobile/package.json
@@ -40,6 +40,7 @@
     "react-native-screens": "~3.22.0",
     "react-native-switch": "^1.5.1",
     "react-native-tab-view": "^3.5.2",
+    "react-native-video": "^5.2.1",
     "react-native-web": "~0.19.6",
     "react-redux": "^8.1.3"
   },
